Scuffed combat

+++ Baj Baj Castle/Assets/Scripts/Combat/DamageData.cs
@@ -1,11 +1,11 @@
 public class DamageData
 {
-    public int Amount;
+    public float Amount;
     public DamageType Type;
     public float Knockback;
     public Actor Source;
 
-    public DamageData(int amount, DamageType type, float knockback, Actor source)
+    public DamageData(float amount, DamageType type, float knockback, Actor source)
     {
         Amount = amount;
         Type = type;
